Force schemas with no properties to be mixed because there is nothing to hydrate into them
