+++ src/Gatherer/OperationHydrator.php
@@ -2,16 +2,6 @@
 
 namespace ApiClients\Tools\OpenApiClientGenerator\Gatherer;
 
-use ApiClients\Tools\OpenApiClientGenerator\Utils;
-use ApiClients\Tools\OpenApiClientGenerator\Registry\Schema as SchemaRegistry;
-use ApiClients\Tools\OpenApiClientGenerator\Representation\OperationRequestBody;
-use ApiClients\Tools\OpenApiClientGenerator\Representation\OperationResponse;
-use ApiClients\Tools\OpenApiClientGenerator\Representation\Parameter;
-use cebe\openapi\spec\Operation as openAPIOperation;
-use cebe\openapi\spec\PathItem;
-use Jawira\CaseConverter\Convert;
-use Psr\Http\Message\ResponseInterface;
-
 final class OperationHydrator
 {
     public static function gather(
@@ -22,12 +12,15 @@ final class OperationHydrator
 
         foreach ($operations as $operation) {
             foreach ($operation->response as $response) {
-                $schemaClasses[] = $response->schema;
+                foreach (HydratorUtils::listSchemas($response->schema) as $schema) {
+                    $schemaClasses[] = $schema;
+                }
             }
         }
 
         return Hydrator::gather(
             'Operation\\' . $className,
+            '🌀',
             ...$schemaClasses,
         );
     }
