+++ src/Gatherer/WebHookHydrator.php
@@ -23,7 +23,7 @@ final class WebHookHydrator
         $schemaClasses = [];
         foreach ($webHooks as $webHook) {
             foreach ($webHook->schema as $webHookSchema) {
-                foreach (self::listSchemas($webHookSchema) as $schema) {
+                foreach (HydratorUtils::listSchemas($webHookSchema) as $schema) {
                     $schemaClasses[] = $schema;
                 }
             }
@@ -31,30 +31,8 @@ final class WebHookHydrator
 
         return Hydrator::gather(
             'WebHook\\' . Utils::className($event),
+            '🪝',
             ...$schemaClasses,
         );
-    }
-
-    /**
-     * @return iterable<\ApiClients\Tools\OpenApiClientGenerator\Representation\Schema>
-     */
-    private static function listSchemas(\ApiClients\Tools\OpenApiClientGenerator\Representation\Schema $schema): iterable
-    {
-        yield $schema;
-        foreach ($schema->properties as $property) {
-            foreach ($property->type as $propertyType) {
-                yield from self::listSchemasFromPropertyType($propertyType);
-            }
-        }
-    }
-
-    private static function listSchemasFromPropertyType(PropertyType $propertyType)
-    {
-        if ($propertyType->payload instanceof \ApiClients\Tools\OpenApiClientGenerator\Representation\Schema) {
-            yield $propertyType->payload;
-            yield from self::listSchemas($propertyType->payload);
-        } else if ($propertyType->payload instanceof PropertyType) {
-            yield from self::listSchemasFromPropertyType($propertyType->payload);
-        }
     }
 }
